Pull new/deleted post subscription code into own file
