+++ packages/shared/src/store/useChannelPosts/index.ts
@@ -1,1 +1,6 @@
 export * from './useChannelPosts';
+export {
+  addToChannelPosts,
+  deleteFromChannelPosts,
+  rollbackDeletedChannelPost,
+} from './subscriptions';
